[FFG Econ] Remove 'Pools' Section
Summary
* Remove pools section
* Add pools to future work
* Tweak wording on protocol utility function future work.

The section on Pools--while informative--seems out of scope at the
current state. It does not have any economic analyses but rather
enumerates the types of pools that may exist. While each of the other
sections have parallel structure that is cogent (as described in the
abstract), the section on pools leaves the reader wondering what the
economic takeaway might be.

Acknowledging the overhead (without sharding) and the existence of pools
is helpful, but think that this should be left out as future work.
Analyzing pools, oligopolistic competition can be separate (or
alternatively can more substantially affect sections 2-3.